Add error checks

+++ scr/main.py
@@ -6,6 +6,8 @@ import os, shutil
 import sys
 import time
 
+NO_SELECTED_FILE = ValueError("No selected file found")
+
 def clean():
     os.system("clear")
 
@@ -27,7 +29,8 @@ class Explorer:
         clean() 
 
         self.files = os.listdir(self.working_dir)
-
+        self.selected_file = None
+        
         for i, file in enumerate(self.files):
             if i == self.selected_index:
                 abs_file = os.path.join(self.working_dir, file)
@@ -49,6 +52,9 @@ class Explorer:
         self.update()
 
     def delete_selected(self):
+        if not self.selected_file:
+            raise NO_SELECTED_FILE 
+
         if self.is_dir_selected:
             shutil.rmtree(self.selected_file)
         else:
@@ -57,6 +63,9 @@ class Explorer:
         self.update()
 
     def open_selected(self):
+        if not self.selected_file:
+            raise NO_SELECTED_FILE
+
         if self.is_dir_selected:
             self.working_dir = self.selected_file
             self.selected_index = 0
@@ -90,15 +99,17 @@ class Explorer:
                     case "k":
                         self.move_focus_up()
                     case "d":
+                        if not self.selected_file:
+                            continue
+
                         self.delete_selected()
                     case " ":
+                        if not self.selected_file:
+                            continue
+
                         self.open_selected()
                     case "-":
                         self.dir_up()
-
-                if c == "q":
-                    clean()
-                    break
 
         self.kb.set_normal_term()
